Rename truffle-compile-vyper to @truffle/compile-vyper
