+++ packages/compile-vyper/README.md
@@ -1,2 +1,2 @@
-# truffle-compile-vyper
+# @truffle/compile-vyper
 Vyper compiler interface for truffle.
