+++ packages/compile-vyper/test/sources/SolidityContract.sol
