+++ packages/compile-vyper/test/sources/VyperContract1.vy
