+++ packages/compile-vyper/test/sources/VyperContract2.v.py
