+++ packages/compile-vyper/test/sources/VyperContract3.vyper.py
