+++ packages/workflow-compile/legacy/index.js
@@ -2,7 +2,7 @@ const debug = require("debug")("workflow-compile");
 const mkdirp = require("mkdirp");
 const { promisify } = require("util");
 const solcCompile = require("@truffle/compile/legacy");
-const vyperCompile = require("truffle-compile-vyper");
+const vyperCompile = require("@truffle/compile-vyper");
 const externalCompile = require("truffle-external-compile");
 const { prepareConfig, multiPromisify } = require("../utils");
 const {
