+++ packages/workflow-compile/package.json
@@ -7,7 +7,7 @@
     "mkdirp": "^0.5.1",
     "truffle-artifactor": "^4.0.29",
     "@truffle/compile": "^4.2.2",
-    "truffle-compile-vyper": "^1.0.26",
+    "@truffle/compile-vyper": "^1.0.26",
     "truffle-config": "^1.1.19",
     "@truffle/expect": "^0.0.9",
     "truffle-external-compile": "^1.0.15",
